nodeAuth
What I think is going on with the DB is caused by using the database before properly configuring it 
to operate. I created (somehow) dbs and collections without configuring an admin account and 
without linking /data/db/ in the /etc/mongod.conf.
It has a style problem also. I am struggling a bit with this. Docker being pitched as much easier is not 
helping. LoL